Moved indexTest.html and indexTest.css to src/assets folder; Created Components and Pages files; Updated App.js; Updtaed index.css. Note: this all need s to be checked...probably a lot of corrections to be made. The goal here is to get this code up and running on the client side. Nothing done yet in src/utils. Resolved intial errors. Webpack is compiling successfully. Errors now on console need to be resolved.
